null case handle

+++ src/main/java/ru/spbau/mit/DictionaryImpl.java
@@ -24,11 +24,24 @@ public class DictionaryImpl implements Dictionary {
 
     @Override
     public boolean contains(String key) {
-        return get(key) != null;
+        if (key == null) {
+            return false;
+        }
+
+        int hash = getHash(key);
+        if (buckets[hash] == null) {
+            return false;
+        }
+        return buckets[hash].contains(key);
     }
 
     @Override
     public String get(String key) {
+
+        if (key == null) {
+            return null;
+        }
+
         int hash = getHash(key);
         if (buckets[hash] == null) {
             return null;
@@ -38,37 +51,49 @@ public class DictionaryImpl implements Dictionary {
 
     @Override
     public String put(String key, String value) {
+
+        if (key == null) {
+            return null;
+        }
+
         int hash = getHash(key);
         if (buckets[hash] == null) {
             buckets[hash] = new List();
         }
         List bucket = buckets[hash];
+
+        if (!bucket.contains(key)) {
+            size++;
+        }
+
         String prevVal = bucket.remove(key);
         bucket.push(key, value);
 
-        if (prevVal == null) {
-            size++;
-            if (size > capacity * (2 + 1) / (2 + 2)) {
-                rehash();
-            }
+        if (size > capacity * (2 + 1) / (2 + 2)) { // 3 / 4
+            rehash();
         }
+
         return prevVal;
     }
 
     @Override
     public String remove(String key) {
+
+        if (key == null) {
+            return null;
+        }
+
         int hash = getHash(key);
         if (buckets[hash] == null) {
             return null;
         }
 
-
         List bucket = buckets[hash];
-        String removed = bucket.remove(key);
-        if (removed != null) {
+        if (bucket.contains(key)) {
             size--;
         }
-        return removed;
+
+        return bucket.remove(key);
     }
 
     @Override
@@ -133,6 +158,15 @@ public class DictionaryImpl implements Dictionary {
                 }
             }
             return null;
+        }
+
+        boolean contains(String key) {
+            for (Node cur = root; cur != null; cur = cur.next) {
+                if (cur.key.equals(key)) {
+                    return true;
+                }
+            }
+            return false;
         }
 
         class Node {
